move kak to quantum_info. add Unitary definition
